refactor: rename base concept file and introduce base named concept
