--- src/data/util/baseConcept.ts
@@ -1,3 +1,0 @@
-export default interface BaseConcept {
-  id: string
-}
