+++ src/data/util/baseConceptTypes.ts
@@ -1,0 +1,5 @@
+export interface BaseConcept {
+  id: string
+}
+
+export type BaseNamedConcept = BaseConcept & { name: string }
